Fixed -generateBuildFile and -generateAllBuild
The original version was too greedy and added all plugin dependencies,
independently if they were "visible" through MANIFEST.MF of given plugin
or not.

This was known limitation - it exposed too much dependencies to compiler
but it worked on Java 8 because on Java 8 compiler didn't care about
duplicated packages. On Java 9+ this is not possible anymore, because at
compile time same package that is contributed by different modules is
not allowed.

https://stackoverflow.com/questions/51094274/eclipse-cant-find-xml-related-classes-after-switching-build-path-to-jdk-10/53824670#53824670

Also added "Reexported by" and "Compilation dependencies" nodes on the
UI.